requirements.txt
pandas
python-docx
torch==2.0.0
torchvision==0.15.1
modelscope
transformers
xtuner[all]
lmdeploy[all]==0.2.4
streamlit
lagent==0.2.2
onnxruntime-gpu
openxlab
--index-url https://download.pytorch.org/whl/cu118 # https://github.com/InternLM/lmdeploy/issues/1169